Generator/Iotsensor/requirements.txt
google-cloud-pubsub
googlemaps:latest